Refactor search logic in Notesbar component and search screen component
