+++ packages/app-mobile/components/NotesBar.tsx
@@ -9,8 +9,7 @@ import { Style } from './global-style';
 import Note from '@joplin/lib/models/Note';
 import NotesBarListItem from './NotesBarListItem';
 import Folder from '@joplin/lib/models/Folder';
-import SearchEngineUtils from '@joplin/lib/services/searchengine/SearchEngineUtils';
-import SearchEngine from '@joplin/lib/services/searchengine/SearchEngine';
+import HandleNoteQuery from './HandleNoteQuery';
 
 interface Props {
     themeId: number;
@@ -190,35 +189,8 @@ function NotesBarComponent(props: Props) {
 
 	// Copied from './screens/search.js' and modified
 	const refreshSearch = async () => {
-		let notes_ = [];
-
-		if (query) {
-			if (props.settings['db.ftsEnabled']) {
-				notes_ = await SearchEngineUtils.notesForQuery(query, true);
-			} else {
-				const p = query.split(' ');
-				const temp = [];
-				for (let i = 0; i < p.length; i++) {
-					const t = p[i].trim();
-					if (!t) continue;
-					temp.push(t);
-				}
-
-				notes_ = await Note.previews(null, {
-					anywherePattern: `*${temp.join('*')}*`,
-				});
-			}
-
-			const parsedQuery = await SearchEngine.instance().parseQuery(query);
-			const highlightedWords = SearchEngine.instance().allParsedQueryTerms(parsedQuery);
-
-			props.dispatch({
-				type: 'SET_HIGHLIGHTED',
-				words: highlightedWords,
-			});
-
-			setNotes(notes_);
-		}
+		const notes = await HandleNoteQuery(query, props.settings, props.dispatch);
+		setNotes(notes);
 	};
 
 	const handleQuerySubmit = async () => {
