+++ packages/app-mobile/components/screens/search.js
@@ -5,13 +5,12 @@ const { connect } = require('react-redux');
 const { ScreenHeader } = require('../ScreenHeader');
 const Icon = require('react-native-vector-icons/Ionicons').default;
 const { _ } = require('@joplin/lib/locale');
-const Note = require('@joplin/lib/models/Note').default;
 const { NoteItem } = require('../note-item.js');
 const { BaseScreenComponent } = require('../base-screen.js');
 const { themeStyle } = require('../global-style.js');
 const DialogBox = require('react-native-dialogbox').default;
-const SearchEngineUtils = require('@joplin/lib/services/searchengine/SearchEngineUtils').default;
 const SearchEngine = require('@joplin/lib/services/searchengine/SearchEngine').default;
+import HandleNoteQuery from '../HandleNoteQuery';
 
 Icon.loadFont();
 
@@ -101,25 +100,7 @@ class SearchScreenComponent extends BaseScreenComponent {
 
 		query = query === null ? this.state.query.trim : query.trim();
 
-		let notes = [];
-
-		if (query) {
-			if (this.props.settings['db.ftsEnabled']) {
-				notes = await SearchEngineUtils.notesForQuery(query, true);
-			} else {
-				const p = query.split(' ');
-				const temp = [];
-				for (let i = 0; i < p.length; i++) {
-					const t = p[i].trim();
-					if (!t) continue;
-					temp.push(t);
-				}
-
-				notes = await Note.previews(null, {
-					anywherePattern: `*${temp.join('*')}*`,
-				});
-			}
-		}
+		const notes = await HandleNoteQuery(query, this.props.settings, this.props.dispatch);
 
 		if (!this.isMounted_) return;
